--- a/js/libdeno.ts
+++ b/js/libdeno.ts
@@ -10,11 +10,15 @@ export type PromiseRejectEvent =
   | "RejectAfterResolved";
 
 interface Libdeno {
+  // TODO Remove recv(), it is unused.
   recv(cb: MessageCallback): void;
 
   send(control: ArrayBufferView, data?: ArrayBufferView): null | Uint8Array;
 
   print(x: string, isErr?: boolean): void;
+
+  runMicrotasks(): void;
+  exit(code: number): void;
 
   setGlobalErrorHandler: (
     handler: (
@@ -34,8 +38,6 @@ interface Libdeno {
       promise: Promise<any>
     ) => void
   ) => void;
-
-  setPromiseErrorExaminer: (handler: () => boolean) => void;
 
   mainSource: string;
   mainSourceMap: RawSourceMap;