--- a/src/libdeno.rs
+++ b/src/libdeno.rs
@@ -33,7 +33,7 @@ extern "C" {
   pub fn deno_new(cb: DenoRecvCb) -> *const isolate;
   pub fn deno_delete(i: *const isolate);
   pub fn deno_last_exception(i: *const isolate) -> *const c_char;
-  pub fn deno_check_promise_errors(i: *const isolate);
+  pub fn deno_run_microtasks(i: *const isolate);
   pub fn deno_respond(
     i: *const isolate,
     user_data: *mut c_void,