@ry 2018-10-17
Before:
```
gpu ~/src/deno> ./tools/throughput_benchmark.py out/before/deno 5000
listening on 127.0.0.1:4544
head -c 5242880000 /dev/zero | nc 127.0.0.1 4544
34.2299828529 seconds
```
After:
```
gpu ~/src/deno> ./tools/throughput_benchmark.py out/after/deno 5000
listening on 127.0.0.1:4544
head -c 5242880000 /dev/zero | nc 127.0.0.1 4544
18.5639469624 seconds
```


Running `strace -c ./out/release/deno tests/fetch_deps.ts --allow-net`

Before this patch:
```
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 99.90    0.003860         161        24           futex
  0.10    0.000004           0       634           mprotect
  0.00    0.000000           0        22           read
  0.00    0.000000           0         3           write
  0.00    0.000000           0        16           open
  0.00    0.000000           0        19           close
  0.00    0.000000           0         2           stat
  0.00    0.000000           0        15           fstat
  0.00    0.000000           0        60           mmap
  0.00    0.000000           0        63           munmap
  0.00    0.000000           0         2           brk
  0.00    0.000000           0         5           rt_sigaction
  0.00    0.000000           0         1           rt_sigprocmask
  0.00    0.000000           0         5         1 ioctl
  0.00    0.000000           0         9         9 access
  0.00    0.000000           0        34        17 madvise
  0.00    0.000000           0         7           clone
  0.00    0.000000           0         1           execve
  0.00    0.000000           0         1           getcwd
  0.00    0.000000           0         2         2 mkdir
  0.00    0.000000           0         1         1 readlink
  0.00    0.000000           0         2           getrlimit
  0.00    0.000000           0         3           sigaltstack
  0.00    0.000000           0         1           arch_prctl
  0.00    0.000000           0         4           sched_getaffinity
  0.00    0.000000           0         1           set_tid_address
  0.00    0.000000           0         1           epoll_ctl
  0.00    0.000000           0         1           set_robust_list
  0.00    0.000000           0         1           epoll_create1
  0.00    0.000000           0         1           pipe2
  0.00    0.000000           0         4           getrandom
------ ----------- ----------- --------- --------- ----------------
100.00    0.003864                   945        30 total
```

After:
```
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
  0.00    0.000000           0        22           read
  0.00    0.000000           0         3           write
  0.00    0.000000           0        16           open
  0.00    0.000000           0        19           close
  0.00    0.000000           0         2           stat
  0.00    0.000000           0        15           fstat
  0.00    0.000000           0        59           mmap
  0.00    0.000000           0       634           mprotect
  0.00    0.000000           0        59           munmap
  0.00    0.000000           0         2           brk
  0.00    0.000000           0         5           rt_sigaction
  0.00    0.000000           0         1           rt_sigprocmask
  0.00    0.000000           0         5         1 ioctl
  0.00    0.000000           0         9         9 access
  0.00    0.000000           0        34        17 madvise
  0.00    0.000000           0         7           clone
  0.00    0.000000           0         1           execve
  0.00    0.000000           0         1           getcwd
  0.00    0.000000           0         2         2 mkdir
  0.00    0.000000           0         1         1 readlink
  0.00    0.000000           0         2           getrlimit
  0.00    0.000000           0         3           sigaltstack
  0.00    0.000000           0         1           arch_prctl
  0.00    0.000000           0        23           futex
  0.00    0.000000           0         4           sched_getaffinity
  0.00    0.000000           0         1           set_tid_address
  0.00    0.000000           0         1           epoll_ctl
  0.00    0.000000           0         1           set_robust_list
  0.00    0.000000           0         1           epoll_create1
  0.00    0.000000           0         1           pipe2
  0.00    0.000000           0         2           getrandom
------ ----------- ----------- --------- --------- ----------------
100.00    0.000000                   937        30 total
```