@kevinkassimo 2018-10-17
Probably we need a benchmark dedicated for futex. Looks pretty bad