@ry 2018-10-17
@kevinkassimo futex is most likely from [the std::sync::mpsc we use in the main thread](https://github.com/denoland/deno/blob/32f07971284f05993cf702aa589859e7900406b4/src/isolate.rs#L223-L227). It looks bad, and I'm not sure it's the best design, but I doubt this is the source of slowness. (I'd rather do epoll the main thread - but it will require [making a custom Tokio runtime](https://tokio.rs/docs/going-deeper/building-runtime/).)
Likely the source of slowness is from flatbuffer serialization on the JS side. It allocates a new ArrayBuffer for each message, and this patch makes each poll serialize a message. I'm going to try to validate this hypothesis with the profiler before trying to fix it. More updates soon...