@ry 2018-10-17
FYI Here's how I am profiling:
```
# Make sure we're only building release.
> export DENO_BUILD_MODE=release
# Build deno and V8's d8.
> ./tools/build.py d8 deno
# Start the program we want to benchmark with --prof
> ./out/release/deno tests/http_bench.ts --allow-net --prof &
# Exercise it.
> third_party/wrk/linux/wrk http://localhost:4500/
# Kill server.
> kill `pgrep deno`
# When supplying --prof, V8 will write a file in the current directory that looks 
# like this isolate-0x7fad98242400-v8.log
# To examine this file
> D8_PATH=out/release/ ./third_party/v8/tools/linux-tick-processor isolate-0x7fad98242400-v8.log
```

This produces a profile: https://gist.github.com/ry/65c9e620225f3d1829cfe39e88cb8467