@ry 2018-10-17
Counter-intuitively this patch makes the program spend more time in C++. Look at the summary:

```
# BEFORE
 [Summary]:
   ticks  total  nonlib   name
   1780   20.3%   21.1%  JavaScript
   6606   75.3%   78.2%  C++
    373    4.3%    4.4%  GC
    325    3.7%          Shared libraries
     62    0.7%          Unaccounted

# AFTER
 [Summary]:
   ticks  total  nonlib   name
    985    9.6%    9.8%  JavaScript
   8981   87.1%   89.7%  C++
    391    3.8%    3.9%  GC
    296    2.9%          Shared libraries
     46    0.4%          Unaccounted
```