@ry 2018-10-25
Closing this one. It resulted in profiling and ongoing perf fixes. The ultimate goal is to move the loop to JS - but we're not ready to do it yet.